tests/requirements.txt
# Test Dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
pytest-env==1.1.3
pytest-timeout==2.2.0
uvicorn==0.27.0

# HTTP Testing
httpx==0.26.0
requests==2.31.0
requests-mock==1.11.0

# Database Testing
faker==22.0.0
factory-boy==3.3.0

# Auth Service Dependencies
fastapi==0.109.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
argon2-cffi==23.1.0
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Utilities
freezegun==1.4.0  # Time mocking

